+++ src/middleware.ts
@@ -1,4 +1,4 @@
-import { stackMiddlewares } from "./lib/middlewares/stackMiddlewares";
+import { stackMiddlewares } from "./middlewares/stackMiddlewares";
 import withUrlHeader from "./middlewares/withUrlHeader";
 
 export default stackMiddlewares([withUrlHeader]);
